requirements.txt
fastapi
uvicorn
python-multipart
pillow
transformers
torch
requests
openai
